progress on test consolidation and cleanup
