+++ __tests__/pages/api/index.test.ts
@@ -1,15 +1,6 @@
-// noinspection DuplicatedCode
-
-import fs from 'fs'
-import Joi from 'joi'
-import YAML from 'yaml'
-import { getTranslations, Translations } from '../../../i18n/api'
-import { BenefitHandler } from '../../../utils/api/benefitHandler'
-import { countryList } from '../../../utils/api/definitions/countries'
 import {
   EntitlementResultType,
   EstimationSummaryState,
-  Language,
   LegalStatus,
   LivingCountry,
   MaritalStatus,
@@ -17,1155 +8,10 @@ import {
   ResultKey,
   ResultReason,
 } from '../../../utils/api/definitions/enums'
-import {
-  FieldDataDropdown,
-  fieldDefinitions,
-  FieldKey,
-} from '../../../utils/api/definitions/fields'
-import { RequestSchema } from '../../../utils/api/definitions/schemas'
-import { OutputItem } from '../../../utils/api/scrapers/_baseTable'
-import { legalValues, scraperData } from '../../../utils/api/scrapers/output'
-import {
-  mockGetRequest,
-  mockGetRequestError,
-  mockPartialGetRequest,
-} from './factory'
+import { legalValues } from '../../../utils/api/scrapers/output'
+import { mockGetRequest, mockPartialGetRequest } from './factory'
 
-describe('code checks', () => {
-  it('produces a list of fields with unique ordering', async () => {
-    const ordersOrig = []
-    for (const key in fieldDefinitions) {
-      ordersOrig.push(fieldDefinitions[key].order)
-    }
-    const ordersUnique = [...new Set(ordersOrig)]
-    expect(ordersUnique).toEqual(ordersOrig)
-  })
-})
-
-describe('translation checks', () => {
-  it('matches between question translations and available questions', async () => {
-    const translationsEn: Translations = getTranslations(Language.EN)
-    const translationsFr: Translations = getTranslations(Language.FR)
-    const translationKeysEn = Object.keys(translationsEn.question)
-    const translationKeysFr = Object.keys(translationsFr.question)
-    const enumKeys = Object.values(FieldKey)
-    expect(translationKeysEn).toEqual(enumKeys)
-    expect(translationKeysFr).toEqual(enumKeys)
-  })
-})
-
-describe('country checks', () => {
-  const COUNTRY_COUNT = 195
-  const handlerEn = new BenefitHandler({ _language: Language.EN })
-  handlerEn.requiredFields = [FieldKey.LIVING_COUNTRY]
-  const fieldDataEn = handlerEn.fieldData as Array<FieldDataDropdown>
-  const handlerFr = new BenefitHandler({ _language: Language.FR })
-  handlerFr.requiredFields = [FieldKey.LIVING_COUNTRY]
-  const fieldDataFr = handlerFr.fieldData as Array<FieldDataDropdown>
-  it(`produces a list of ${COUNTRY_COUNT} countries (EN and FR)`, async () => {
-    expect(fieldDataEn[0].values.length).toEqual(COUNTRY_COUNT)
-    expect(fieldDataFr[0].values.length).toEqual(COUNTRY_COUNT)
-  })
-  it(`produces a list of countries with Canada first and AFG second (EN and FR)`, async () => {
-    expect(fieldDataEn[0].values[0].key).toEqual('CAN') // ensure Canada is first in the list
-    expect(fieldDataEn[0].values[1].key).toEqual('AFG') // ensure Agreement is not in the list (AFG should be next)
-    expect(fieldDataFr[0].values[0].key).toEqual('CAN')
-    expect(fieldDataFr[0].values[1].key).toEqual('AFG')
-  })
-  it(`includes Agreement in "agreement countries" list`, async () => {
-    expect(countryList[0].code).toEqual('CAN') // ensure Canada is first in the list
-    expect(countryList[1].code).toEqual('AGREEMENT') // ensure Agreement is second in the list
-  })
-})
-
-describe('schema checks', () => {
-  function getJoiKeys(schema: Joi.ObjectSchema) {
-    // @ts-ignore
-    const joiEntries = schema._ids._byKey.entries()
-    const joiKeys = []
-    for (const joiEntry of joiEntries) {
-      if (joiEntry[0][0] === '_') continue // ignore system fields
-      joiKeys.push(joiEntry[0])
-    }
-    return joiKeys
-  }
-
-  it('matches between field definitions and schema', async () => {
-    const joiKeys = getJoiKeys(RequestSchema)
-    const enumKeys = Object.values(FieldKey)
-    expect(joiKeys).toEqual(enumKeys)
-  })
-})
-
-describe('openapi checks', () => {
-  const file = fs.readFileSync('./public/openapi.yaml', 'utf-8')
-  const openapi = YAML.parse(file)
-  it('matches LegalStatus enum', async () => {
-    expect(openapi.components.parameters.legalStatus.schema.enum).toEqual(
-      Object.values(LegalStatus)
-    )
-  })
-  it('matches MaritalStatus enum', async () => {
-    expect(openapi.components.parameters.maritalStatus.schema.enum).toEqual(
-      Object.values(MaritalStatus)
-    )
-  })
-  it('matches Country enum', async () => {
-    expect(openapi.components.parameters.livingCountry.schema.enum).toEqual(
-      Object.values(LivingCountry)
-    )
-  })
-  it('matches FieldKey enum', async () => {
-    expect(openapi.components.schemas.FieldKey.items.enum).toEqual(
-      Object.values(FieldKey)
-    )
-  })
-  it('matches ResultKey enum', async () => {
-    expect(openapi.components.schemas.ResultKey.enum).toEqual(
-      Object.values(ResultKey)
-    )
-  })
-  it('matches ResultReason enum', async () => {
-    expect(openapi.components.schemas.ResultReason.enum).toEqual(
-      Object.values(ResultReason)
-    )
-  })
-  it('matches parameters', async () => {
-    const openApiParams = Object.keys(openapi.components.parameters)
-    const enumKeys = Object.values(FieldKey)
-    expect(openApiParams).toEqual(Object.values(enumKeys))
-    const openApiPathParams =
-      openapi.paths['/calculateEligibility'].get.parameters
-    const openApiPathParamsStripped = openApiPathParams.map((x) =>
-      x['$ref'].replace('#/components/parameters/', '')
-    )
-    expect(openApiPathParamsStripped).toEqual(Object.values(enumKeys))
-  })
-})
-
-describe('scraper tests', () => {
-  it("the scraped data's last entry is less than one", async () => {
-    const toVerify: { data: OutputItem[]; key: string }[] = [
-      { data: scraperData.tbl1_single, key: 'gis' },
-      { data: scraperData.tbl2_partneredAndOas, key: 'gis' },
-      { data: scraperData.tbl3_partneredNoOas, key: 'gis' },
-      { data: scraperData.tbl4_partneredAlw, key: 'alw' },
-      { data: scraperData.tbl5_partneredAfs, key: 'afs' },
-    ]
-    toVerify.forEach((value) => {
-      const last = value.data[value.data.length - 1][value.key]
-      expect(last).toBeLessThan(1)
-    })
-  })
-})
-
-describe('sanity checks', () => {
-  it('fails on income with letters', async () => {
-    const res = await mockGetRequestError({
-      income: 'abc' as unknown as number,
-    })
-    expect(res.status).toEqual(400)
-    expect(res.body.error).toEqual(ResultKey.INVALID)
-  })
-  it('fails on age over 150', async () => {
-    const res = await mockGetRequestError({ age: 151 })
-    expect(res.status).toEqual(400)
-    expect(res.body.error).toEqual(ResultKey.INVALID)
-  })
-  it('accepts age equal to 150', async () => {
-    const res = await mockPartialGetRequest({ age: 150 })
-    expect(res.status).toEqual(200)
-  })
-  it('accepts valid Marital Status', async () => {
-    const res = await mockPartialGetRequest({
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-    })
-    expect(res.status).toEqual(200)
-  })
-  it('accepts valid Legal Status', async () => {
-    const res = await mockPartialGetRequest({
-      age: 65,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-    })
-    expect(res.status).toEqual(200)
-  })
-  it('fails when years in Canada is greater than age minus 18', async () => {
-    const res = await mockGetRequestError({
-      age: 65,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 48,
-    })
-    expect(res.status).toEqual(400)
-    expect(res.body.error).toEqual(ResultKey.INVALID)
-  })
-  it('accepts when years in Canada is equal to age minus 18', async () => {
-    const res = await mockPartialGetRequest({
-      age: 65,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 47,
-    })
-    expect(res.status).toEqual(200)
-  })
-})
-
-describe('field requirement analysis', () => {
-  it('requires only income when nothing provided', async () => {
-    const res = await mockGetRequest({
-      income: undefined,
-      age: undefined,
-      maritalStatus: undefined,
-      livingCountry: undefined,
-      legalStatus: undefined,
-      canadaWholeLife: undefined,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-    ])
-  })
-  it('requires fields when only income provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: undefined,
-      maritalStatus: undefined,
-      livingCountry: undefined,
-      legalStatus: undefined,
-      canadaWholeLife: undefined,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.MARITAL_STATUS,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-    ])
-  })
-  it('requires fields when only income/age provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: undefined,
-      livingCountry: undefined,
-      legalStatus: undefined,
-      canadaWholeLife: undefined,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.MARITAL_STATUS,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-    ])
-  })
-  it('requires fields when only income/age/marital provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: undefined,
-      legalStatus: undefined,
-      canadaWholeLife: undefined,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-  it('requires fields when only income/age/marital/country provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: undefined,
-      canadaWholeLife: undefined,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-  it('requires fields when only income/age/marital/country/legal provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: undefined,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-  it('requires fields when only income/age/marital/country/legal/lifeCanada provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-  it('requires fields when only income/age/marital/country/legal/lifeCanada/years provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 5,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.EVER_LIVED_SOCIAL_COUNTRY,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.EVER_LIVED_SOCIAL_COUNTRY,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-  it('requires fields when only income/age/marital/country/legal/lifeCanada/years/socialCountry provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 5,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.EVER_LIVED_SOCIAL_COUNTRY,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-  it('requires fields when only income/age/marital/country/legal/lifeCanada/years/socialCountry/partnerBenefits provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 5,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([FieldKey.PARTNER_INCOME])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.EVER_LIVED_SOCIAL_COUNTRY,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-  it('requires no fields when all provided', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 5,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 10000,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.UNAVAILABLE)
-    expect(res.body.missingFields).toEqual([])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.EVER_LIVED_SOCIAL_COUNTRY,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-})
-
-describe('field requirements analysis: conditional fields', () => {
-  it('requires "yearsInCanadaSince18" when lifeCanada=false', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([FieldKey.YEARS_IN_CANADA_SINCE_18])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-    ])
-  })
-  it('requires "everLivedSocialCountry" when citizen and under 10 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 5,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([FieldKey.EVER_LIVED_SOCIAL_COUNTRY])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.YEARS_IN_CANADA_SINCE_18,
-      FieldKey.EVER_LIVED_SOCIAL_COUNTRY,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-    ])
-  })
-  it('requires partner questions when marital=married', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-    expect(res.body.missingFields).toEqual([
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-    expect(res.body.visibleFields).toEqual([
-      FieldKey.AGE,
-      FieldKey.LIVING_COUNTRY,
-      FieldKey.LEGAL_STATUS,
-      FieldKey.CANADA_WHOLE_LIFE,
-      FieldKey.INCOME,
-      FieldKey.MARITAL_STATUS,
-      FieldKey.PARTNER_BENEFIT_STATUS,
-      FieldKey.PARTNER_INCOME,
-    ])
-  })
-})
-
-describe('summary object checks', () => {
-  it('returns "available eligible"', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 10000,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(
-      EstimationSummaryState.AVAILABLE_ELIGIBLE
-    )
-  })
-  it('returns "available ineligible"', async () => {
-    const res = await mockGetRequest({
-      income: 20000,
-      age: 50,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 10000,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(
-      EstimationSummaryState.AVAILABLE_INELIGIBLE
-    )
-  })
-  it('returns "unavailable"', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.SPONSORED,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 10000,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.UNAVAILABLE)
-  })
-})
-
-describe('basic OAS scenarios', () => {
-  it(`returns income error when income equal to ${legalValues.MAX_OAS_INCOME}`, async () => {
-    const res = await mockGetRequestError({
-      income: legalValues.MAX_OAS_INCOME,
-    })
-    expect(res.status).toEqual(400)
-    expect(res.body.error).toEqual(ResultKey.INVALID)
-    if (!('details' in res.body.detail)) throw Error('missing details')
-    expect(res.body.detail.details[0].path[0]).toEqual(FieldKey.INCOME)
-  })
-  it('returns "unavailable" when not citizen (other provided)', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.OTHER,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.LEGAL_STATUS
-    )
-  })
-  it('returns "unavailable" when sponsored', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.SPONSORED,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.LEGAL_STATUS
-    )
-  })
-  it('returns "unavailable" when citizen and under 10 years in Canada and lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 9,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-  })
-  it('returns "ineligible" when citizen and under 10 years in Canada and not lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 9,
-      everLivedSocialCountry: false,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-  })
-  it('returns "eligible" when citizen and 10 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 10,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.oas.entitlement.type).toEqual(
-      EntitlementResultType.PARTIAL
-    )
-  })
-  it('returns "eligible" when living in Agreement and 20 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.oas.entitlement.type).toEqual(
-      EntitlementResultType.PARTIAL
-    )
-  })
-  it('returns "unavailable" when living in Agreement and under 20 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 19,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-  })
-  it('returns "eligible" when living in No Agreement and 20 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.NO_AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.oas.entitlement.type).toEqual(
-      EntitlementResultType.PARTIAL
-    )
-  })
-  it('returns "needs more info" when living in No Agreement and under 20 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: undefined,
-      livingCountry: LivingCountry.NO_AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 19,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.summary.state).toEqual(EstimationSummaryState.MORE_INFO)
-  })
-  it('returns "ineligible" when living in No Agreement and under 20 years in Canada and not lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.NO_AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 19,
-      everLivedSocialCountry: false,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-  })
-  it('returns "ineligible due to age" when 64 and 9 years in Canada and lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 64,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 9,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(ResultReason.AGE)
-  })
-  it('returns "unavailable" when living in No Agreement and under 20 years in Canada and lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.NO_AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 19,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-  })
-  it('returns "eligible when 65" when age 55 and citizen and 20 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 55,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.AGE_YOUNG
-    )
-  })
-  it('returns "ineligible" when age 55 and legal=sponsored and 20 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 55,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.SPONSORED,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(ResultReason.AGE)
-  })
-  it('returns "ineligible" when age 55 and legal=other and 20 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 55,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.OTHER,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(ResultReason.AGE)
-  })
-})
+describe('basic OAS scenarios', () => {})
 
 describe('OAS entitlement scenarios', () => {
   it('returns "eligible for $317.63" when 20 years in Canada', async () => {
@@ -1240,260 +86,6 @@ describe('OAS entitlement scenarios', () => {
     expect(res.body.results.oas.eligibility.result).toEqual(ResultKey.ELIGIBLE)
     expect(res.body.results.oas.entitlement.result).toEqual(648.67)
     expect(res.body.results.oas.eligibility.reason).toEqual(ResultReason.NONE)
-  })
-})
-
-describe('basic GIS scenarios', () => {
-  it('returns income error when income 1,000,000', async () => {
-    const res = await mockGetRequestError({
-      maritalStatus: MaritalStatus.MARRIED,
-      income: 1000000,
-      age: 65,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-    })
-    expect(res.status).toEqual(400)
-    expect(res.body.error).toEqual(ResultKey.INVALID)
-    if (!('details' in res.body.detail)) throw Error('missing details')
-    expect(res.body.detail.details[0].path[0]).toEqual(FieldKey.INCOME)
-  })
-  it('returns "ineligible" when not living in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.NO_AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(
-      ResultReason.LIVING_COUNTRY
-    )
-  })
-  it('returns "unavailable" when sponsored', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.SPONSORED,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(
-      ResultReason.LEGAL_STATUS
-    )
-  })
-  it(`returns "ineligible" when single and income equal to ${legalValues.MAX_GIS_INCOME_SINGLE}`, async () => {
-    const res = await mockGetRequest({
-      income: legalValues.MAX_GIS_INCOME_SINGLE,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.INCOME)
-  })
-  it('returns "eligible" when single and income under 19248', async () => {
-    const res = await mockGetRequest({
-      income: 19247,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.NONE)
-  })
-  it(`returns "ineligible" when married and no partner OAS and income equal to ${legalValues.MAX_GIS_INCOME_PARTNER_ALW}`, async () => {
-    const res = await mockGetRequest({
-      income: legalValues.MAX_GIS_INCOME_PARTNER_ALW,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.NONE,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.INCOME)
-  })
-  it('returns "eligible" when married and no partner OAS and income under 45656 and partner income 1000', async () => {
-    const res = await mockGetRequest({
-      income: 45655,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.NONE,
-      partnerIncome: 1000,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.NONE)
-  })
-  it(`returns "ineligible" when married and partner OAS and income equal to ${legalValues.MAX_GIS_INCOME_PARTNER_OAS} and partner income 0`, async () => {
-    const res = await mockGetRequest({
-      income: legalValues.MAX_GIS_INCOME_PARTNER_OAS,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.INCOME)
-  })
-  it(`returns "ineligible" when married and partner OAS and income equal to ${legalValues.MAX_GIS_INCOME_PARTNER_OAS} and partner income 1000`, async () => {
-    const res = await mockGetRequest({
-      income: legalValues.MAX_GIS_INCOME_PARTNER_OAS,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 1000,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.INCOME)
-  })
-  it('returns "eligible" when married and partner OAS and income under 25440 and partner income 0', async () => {
-    const res = await mockGetRequest({
-      income: 25439,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.NONE)
-  })
-  it(`returns "ineligible" when married and partner partial OAS and income equal to ${legalValues.MAX_GIS_INCOME_PARTNER_OAS} and partner income 0`, async () => {
-    const res = await mockGetRequest({
-      income: legalValues.MAX_GIS_INCOME_PARTNER_OAS,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.INCOME)
   })
 })
 
@@ -1743,156 +335,6 @@ describe('GIS entitlement scenarios', () => {
 })
 
 describe('basic Allowance scenarios', () => {
-  it(`returns "ineligible" when income equal to ${legalValues.MAX_ALW_INCOME}`, async () => {
-    const res = await mockGetRequest({
-      income: legalValues.MAX_ALW_INCOME,
-      age: 60,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.INCOME)
-  })
-  it('returns "ineligible due to age" when age 65 and high income', async () => {
-    const res = await mockGetRequest({
-      income: 36049,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.AGE)
-  })
-  it('returns "ineligible" when age over 64', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.AGE)
-  })
-  it('returns "ineligible" when age under 60', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 59,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(
-      ResultReason.AGE_YOUNG
-    )
-  })
-  it('returns "unavailable" when not citizen', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 60,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.OTHER,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 20,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(
-      ResultReason.LEGAL_STATUS
-    )
-  })
-  it('returns "ineligible" when citizen and under 10 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 60,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 9,
-      everLivedSocialCountry: false,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-  })
   it('returns "ineligible" when widowed', async () => {
     const res = await mockGetRequest({
       income: 10000,
@@ -1919,32 +361,7 @@ describe('basic Allowance scenarios', () => {
       ResultReason.MARITAL
     )
   })
-  it('returns "ineligible" when single', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 60,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 10,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(
-      ResultReason.MARITAL
-    )
-  })
+
   it('returns "ineligible" when partner not receiving OAS', async () => {
     const res = await mockGetRequest({
       income: 10000,
@@ -1971,50 +388,7 @@ describe('basic Allowance scenarios', () => {
       ResultReason.PARTNER
     )
   })
-  it('returns "eligible" when citizen and 10 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 60,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 10,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.NONE)
-  })
-  it('returns "eligible" when living in Agreement and 10 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 60,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 10,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.NONE)
-  })
+
   it('returns "unavailable" when living in Agreement and under 10 years in Canada', async () => {
     const res = await mockGetRequest({
       income: 10000,
@@ -2034,12 +408,6 @@ describe('basic Allowance scenarios', () => {
       partnerYearsInCanadaSince18: undefined,
       partnerEverLivedSocialCountry: undefined,
     })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
   })
   it('returns "eligible" when living in No Agreement and 10 years in Canada', async () => {
     const res = await mockGetRequest({
@@ -2089,30 +457,6 @@ describe('basic Allowance scenarios', () => {
       ResultReason.YEARS_IN_CANADA
     )
   })
-  it('returns "ineligible" when under 60, legal=other', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 55,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.OTHER,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 10,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.AGE)
-  })
 })
 
 describe('Allowance entitlement scenarios', () => {
@@ -2139,29 +483,7 @@ describe('Allowance entitlement scenarios', () => {
     expect(res.body.results.alw.entitlement.result).toEqual(341.68)
     expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.NONE)
   })
-  it('returns "eligible for $334.33" when 40 years in Canada and income=20000 and partner=partialOas', async () => {
-    const res = await mockGetRequest({
-      income: 20000,
-      age: 60,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.alw.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.alw.entitlement.result).toEqual(341.68)
-    expect(res.body.results.alw.eligibility.reason).toEqual(ResultReason.NONE)
-  })
+
   it('returns "eligible for $565.35" when 40 years in Canada and income=10000', async () => {
     const res = await mockGetRequest({
       income: 10000,
@@ -2527,7 +849,9 @@ describe('basic Allowance for Survivor scenarios', () => {
     expect(res.body.results.afs.eligibility.result).toEqual(
       ResultKey.INELIGIBLE
     )
-    expect(res.body.results.afs.eligibility.reason).toEqual(ResultReason.AGE)
+    expect(res.body.results.afs.eligibility.reason).toEqual(
+      ResultReason.AGE_YOUNG
+    )
   })
 })
 
@@ -2603,308 +927,7 @@ describe('AFS entitlement scenarios', () => {
   })
 })
 
-describe('thorough personas', () => {
-  it('Tanu Singh: OAS eligible, GIS eligible', async () => {
-    const res = await mockGetRequest({
-      income: 17000,
-      age: 65,
-      maritalStatus: MaritalStatus.MARRIED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: true,
-      yearsInCanadaSince18: undefined,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: PartnerBenefitStatus.OAS_GIS,
-      partnerIncome: 0,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.oas.eligibility.reason).toEqual(ResultReason.NONE)
-    expect(res.body.results.gis.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.NONE)
-  })
-  it('Habon Aden: OAS unavailable, GIS ineligible due to country', async () => {
-    const res = await mockGetRequest({
-      income: 28000,
-      age: 66,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 18,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(
-      ResultReason.LIVING_COUNTRY
-    )
-  })
-  it('Miriam Krayem: OAS eligible when 65, GIS ineligible due to income', async () => {
-    const res = await mockGetRequest({
-      income: 40000,
-      age: 55,
-      maritalStatus: MaritalStatus.DIVORCED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 30,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.AGE_YOUNG
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.OAS)
-  })
-  it('Adam Smith: OAS eligible when 65, GIS ineligible due to income', async () => {
-    const res = await mockGetRequest({
-      income: 25000,
-      age: 62,
-      maritalStatus: MaritalStatus.WIDOWED,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.PERMANENT_RESIDENT,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 15,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.AGE_YOUNG
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.OAS)
-  })
-})
-
 describe('thorough extras', () => {
-  it('returns "ineligible due to years in Canada" when living in Canada and 9 years in Canada and never lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 9,
-      everLivedSocialCountry: false,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.OAS)
-  })
-  it('returns "unavailable" when living in Canada and 9 years in Canada and lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 9,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.OAS)
-  })
-  it('returns "unavailable" when living in Agreement and 9 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 9,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(
-      ResultReason.LIVING_COUNTRY
-    )
-  })
-  it('returns "eligible" when living in Canada and 10 years in Canada', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.CANADA,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 10,
-      everLivedSocialCountry: undefined,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.oas.entitlement.type).toEqual(
-      EntitlementResultType.PARTIAL
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(ResultKey.ELIGIBLE)
-    expect(res.body.results.gis.eligibility.reason).toEqual(ResultReason.NONE)
-  })
-  it('returns "unavailable" when not living in Canada and 19 years in Canada and lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.NO_AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 19,
-      everLivedSocialCountry: true,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.UNAVAILABLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(
-      ResultReason.LIVING_COUNTRY
-    )
-  })
-  it('returns "ineligible due to years in Canada" when not living in Canada and 19 years in Canada and not lived in social country', async () => {
-    const res = await mockGetRequest({
-      income: 10000,
-      age: 65,
-      maritalStatus: MaritalStatus.SINGLE,
-      livingCountry: LivingCountry.NO_AGREEMENT,
-      legalStatus: LegalStatus.CANADIAN_CITIZEN,
-      canadaWholeLife: false,
-      yearsInCanadaSince18: 19,
-      everLivedSocialCountry: false,
-      partnerBenefitStatus: undefined,
-      partnerIncome: undefined,
-      partnerAge: undefined,
-      partnerLivingCountry: undefined,
-      partnerLegalStatus: undefined,
-      partnerCanadaWholeLife: undefined,
-      partnerYearsInCanadaSince18: undefined,
-      partnerEverLivedSocialCountry: undefined,
-    })
-    expect(res.body.results.oas.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.oas.eligibility.reason).toEqual(
-      ResultReason.YEARS_IN_CANADA
-    )
-    expect(res.body.results.gis.eligibility.result).toEqual(
-      ResultKey.INELIGIBLE
-    )
-    expect(res.body.results.gis.eligibility.reason).toEqual(
-      ResultReason.LIVING_COUNTRY
-    )
-  })
   it('returns "eligible" when not living in Canada and 20 years in Canada', async () => {
     const res = await mockGetRequest({
       income: 15000,
