+++ utils/api/benefits/afsBenefit.ts
@@ -91,7 +91,7 @@ export class AfsBenefit extends BaseBenefit {
         } else if (underAgeReq) {
           return {
             result: ResultKey.INELIGIBLE,
-            reason: ResultReason.AGE,
+            reason: ResultReason.AGE_YOUNG,
             detail: this.translations.detail.dependingOnAgreementWhen60,
           }
         } else {
@@ -112,7 +112,7 @@ export class AfsBenefit extends BaseBenefit {
       if (underAgeReq) {
         return {
           result: ResultKey.INELIGIBLE,
-          reason: ResultReason.AGE,
+          reason: ResultReason.AGE_YOUNG,
           detail: this.translations.detail.dependingOnLegalWhen60,
         }
       } else if (this.input.legalStatus.sponsored) {
